# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Active Home Platform API
  version: 1.0.0
  summary: Configuration-driven housing & construction advertising marketplace (Uzbekistan) — v1 contract.
  description: "Enterprise OpenAPI 3.1 contract for **Active Home v1**, the implementation contract between\
    \ the\nweb frontend, backend (FastAPI modular monolith), AI coding agents, QA automation, future mobile\n\
    apps, and external integrations.\n\nThis document is generated from and fully aligned with the immutable\
    \ approved architecture:\nSRS v1.0, DDD Domain Model v1.0, Configuration & Metadata Framework v1.0,\
    \ SAD v1.0,\nDatabase Architecture & Logical Data Model v1.0, Physical Database Design v1.0, and the\n\
    API Architecture & Integration Specification v1.0.\n\n## Cross-cutting contract rules (see the API\
    \ Architecture Specification)\n- **Versioning:** URI major versioning under `/api/v1`. Within a major\
    \ version only additive,\n  backward-compatible change is allowed. Configuration changes (new categories,\
    \ forms, fields,\n  plans, roles, templates) are **never** contract changes — dynamic data flows through\
    \ stable\n  envelopes (`AttributeMap`, `LocalizedText`, `FormDefinition`), so the endpoint surface\
    \ is\n  invariant under the platform's normal mode of change.\n- **Authentication:** server-side sessions\
    \ backed by Redis (SAD §13). A session is established by an\n  auth flow and carried by a secure,\
...
  contact:
...
  license:
...
servers:
...
security:
- sessionCookie: []
...
paths:
...
  /me/preferences:
    put:
      tags:
      - Users
      operationId: updatePreferences
      summary: Update notification & privacy preferences
      description: Updates notification channel opt-ins and phone-reveal privacy (FR-USER-003; OTP always
        via SMS regardless).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                privacySettings:
                  $ref: '#/components/schemas/PrivacySettings'
                notificationPreferences:
                  $ref: '#/components/schemas/NotificationPreferences'
              additionalProperties: false
      responses:
        '200':
          description: Updated account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /reports:
    post:
      tags:
      - Messaging
      operationId: createReport
      summary: Report content
      description: Raises a report against a listing, conversation, or user (FR-MOD-003); routed to a
        moderation case.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '202':
          description: Report received and queued for moderation.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /me/notifications:
    get:
      tags:
      - Notifications
      operationId: listNotifications
      summary: List my notifications
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: unreadOnly
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: A page of notifications.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/CursorPage'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/notifications/{notificationId}/read:
    put:
      tags:
      - Notifications
      operationId: setNotificationRead
      summary: Set notification read status
      parameters:
      - $ref: '#/components/parameters/NotificationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationReadRequest'
      responses:
        '200':
          description: Updated notification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/notifications/read-all:
    post:
      tags:
      - Notifications
      operationId: markAllNotificationsRead
      summary: Mark all notifications read
      responses:
        '204':
          description: All marked read.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/moderation-queue:
    get:
      tags:
      - Administration
      operationId: listModerationQueue
      summary: List moderation cases
      description: The moderation work queue (FR-MOD-005). Filterable by status and subject type.
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - OPEN
          - IN_REVIEW
          - RESOLVED
      - name: subjectType
        in: query
        schema:
          type: string
          enum:
          - LISTING
          - CONVERSATION
          - USER
          - PROFILE
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of moderation cases.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/CursorPage'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/ModerationCase'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/moderation-queue/{caseId}:
    get:
      tags:
      - Administration
      operationId: getModerationCase
      summary: Get a moderation case
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: The case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationCase'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/moderation-queue/{caseId}/action:
    post:
      tags:
      - Administration
      operationId: applyModerationAction
      summary: Resolve a moderation case
      description: 'Applies one verb from the closed set (BR-MOD-02). The action is issued as a command
        to the target

        module (the target owns its state change) and recorded as an immutable audit entry (I-22).

        '
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationActionRequest'
      responses:
        '200':
          description: Case resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationCase'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ah_session
      description: 'First-party, server-side session (Redis-backed, SAD §13). Issued by any auth flow
        and sent as a

        secure, HttpOnly, SameSite=Lax cookie. This is the default scheme for browser clients.

        '
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 'The same first-party session token presented as `Authorization: Bearer <token>` for
        native and

        mobile clients that cannot use cookies. The token is an **opaque server-side session reference**,

        not a JWT. Revoking the session (logout) invalidates it immediately.

        '
    googleOAuth:
      type: oauth2
      description: 'Google is a federated **sign-in** identity provider only (authorization-code flow).
        After Google

        authenticates the user, Active Home links or creates the account and issues its own first-party

        session (`sessionCookie`/`sessionBearer`). Active Home does not expose an OAuth2 authorization

        server of its own.

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect sign-in
...
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque forward cursor from a previous page's `nextCursor`. Omit for the first page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Values above the maximum are clamped, not rejected.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    NotificationIdPath:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: No valid session, or session expired/revoked.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            noSession:
              $ref: '#/components/examples/ErrUnauthorized'
    Forbidden:
      description: Authenticated but not permitted (default-deny; wrong acting profile; insufficient permission
        key).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            denied:
              $ref: '#/components/examples/ErrForbidden'
    NotFound:
      description: Resource does not exist or is not visible to the caller (opaque).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              $ref: '#/components/examples/ErrNotFound'
    Conflict:
      description: Optimistic-concurrency conflict, duplicate unique key, or illegal state transition.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            conflict:
              $ref: '#/components/examples/ErrConflict'
    UnprocessableEntity:
      description: Business-rule violation or field-level validation failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              $ref: '#/components/examples/ErrValidation'
            businessRule:
              $ref: '#/components/examples/ErrBusinessRule'
...
  schemas:
...
    CursorPage:
      type: object
      description: Cursor pagination envelope (opaque forward cursor).
      required:
      - items
      - page
      properties:
        items:
          type: array
          items: {}
        page:
          type: object
          required:
          - limit
          properties:
            limit:
              type: integer
            nextCursor:
              type:
              - string
              - 'null'
              description: Pass as `cursor` to fetch the next page; null when exhausted.
            total:
              type:
              - integer
              - 'null'
              description: Present only where cheap to compute; may be null.
      additionalProperties: false
    Error:
      type: object
      description: 'Single Problem-style error envelope (RFC-9457 shaped). `code` is a stable, machine-readable

        member of a closed vocabulary and refines the HTTP status.

        '
      required:
      - type
      - title
      - status
      - code
      - traceId
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
          examples:
          - https://errors.activehome.uz/validation
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        code:
          type: string
          description: Stable application-level error code.
          enum:
          - VALIDATION_FAILED
          - BUSINESS_RULE_VIOLATION
          - AUTHENTICATION_REQUIRED
          - AUTHENTICATION_INVALID
          - SESSION_EXPIRED
          - PERMISSION_DENIED
          - WRONG_ACTING_PROFILE
          - RESOURCE_NOT_FOUND
          - CONFLICT
          - DUPLICATE_KEY
          - ILLEGAL_STATE_TRANSITION
          - QUOTA_EXCEEDED
          - IMAGE_COUNT_EXCEEDED
          - UNSUPPORTED_MEDIA_TYPE
          - IDEMPOTENCY_KEY_REUSED
          - RATE_LIMITED
          - OTP_THROTTLED
          - DEPENDENCY_DEGRADED
        detail:
          type: string
          description: Human-readable explanation for this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference for this specific occurrence.
        traceId:
          type: string
          description: Correlation id for support and logs.
        details:
          type: array
          description: Field-level errors (validation and business-rule cases).
          items:
            $ref: '#/components/schemas/ValidationError'
      additionalProperties: false
    ValidationError:
      type: object
      required:
      - path
      - rule
      - message
      properties:
        path:
          type: string
          description: JSON pointer / field path.
          examples:
          - /attributes/area_m2
        rule:
          type: string
          description: The whitelisted validator or schema rule that failed.
          examples:
          - numeric_range
        message:
          type: string
      additionalProperties: false
    Account:
      type: object
      description: The authenticated user's own account (BC-01 UserAccount).
      required:
      - id
      - status
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        phoneNumber:
          type:
          - string
          - 'null'
          description: E.164; null after anonymisation.
        email:
          type:
          - string
          - 'null'
          format: email
        displayName:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - CLOSED
          readOnly: true
        privacySettings:
          $ref: '#/components/schemas/PrivacySettings'
        notificationPreferences:
          $ref: '#/components/schemas/NotificationPreferences'
        ownedProfileIds:
          type: array
          items:
            type: string
            format: uuid
          readOnly: true
        roles:
          type: array
          readOnly: true
          items:
            type: string
            description: Role codes assigned to the account.
        createdAt:
          type: string
          format: date-time
          readOnly: true
        accountKind:
          type:
          - string
          - 'null'
          enum:
          - INDIVIDUAL
          - LEGAL_ENTITY
          - INVESTOR
          - null
          readOnly: true
          description: ADR-0007. The role chosen at signup.
        reviewStatus:
          type:
          - string
          - 'null'
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          - null
          readOnly: true
          description: ADR-0007. Gates the role-specific workspace only -- login itself is
            governed by status, unchanged.
        reviewReason:
          type:
          - string
          - 'null'
          readOnly: true
          description: The reviewer's note, once reviewStatus is APPROVED/REJECTED.
      additionalProperties: false
    PrivacySettings:
      type: object
      properties:
        phoneRevealMode:
          type: string
          enum:
          - ALWAYS
          - ON_REQUEST
          - NEVER
          default: ON_REQUEST
          description: BRULE-13 phone-reveal gating.
      additionalProperties: false
    NotificationPreferences:
      type: object
      description: 'Channel opt-ins by category. OTP is always delivered via SMS (Eskiz) regardless of
        preference.

        '
      properties:
        email:
          type: boolean
          default: true
        webPush:
          type: boolean
          default: false
        sms:
          type: boolean
          default: true
      additionalProperties: false
...
    ReportCreateRequest:
      type: object
      required:
      - subjectType
      - subjectId
      - reason
      properties:
        subjectType:
          type: string
          enum:
          - LISTING
          - CONVERSATION
          - USER
          - PROFILE
        subjectId:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 1000
      additionalProperties: false
    Notification:
      type: object
      required:
      - id
      - eventKey
      - channel
      - createdAt
      - deliveryStatus
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        eventKey:
          type: string
          readOnly: true
        channel:
          type: string
          enum:
          - EMAIL
          - WEB_PUSH
          - SMS
          readOnly: true
        subject:
          type:
          - string
          - 'null'
          readOnly: true
        body:
          type: string
          readOnly: true
        readAt:
          type:
          - string
          - 'null'
          format: date-time
        deliveryStatus:
          type: string
          enum:
          - QUEUED
          - SENT
          - DELIVERED
          - FAILED
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
    NotificationReadRequest:
      type: object
      required:
      - read
      properties:
        read:
          type: boolean
      additionalProperties: false
...
    ModerationCase:
      type: object
      required:
      - id
      - subjectType
      - subjectId
      - originType
      - status
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        subjectType:
          type: string
          enum:
          - LISTING
          - CONVERSATION
          - USER
          - PROFILE
        subjectId:
          type: string
          format: uuid
        originType:
          type: string
          enum:
          - USER_REPORT
          - AUTOMATED_FLAG
          readOnly: true
        reportReason:
          type:
          - string
          - 'null'
        ruleKey:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - OPEN
          - IN_REVIEW
          - RESOLVED
          readOnly: true
        resolutionAction:
          type:
          - string
          - 'null'
          enum:
          - HIDE
          - REJECT
          - SUSPEND
          - REQUEST_CORRECTION
          - REMOVE
          - SUSPEND_ACCOUNT
          - DISMISS
          - REVOKE_BADGE
          - ARCHIVE_PROFILE
          - null
        createdAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
    ModerationActionRequest:
      type: object
      required:
      - action
      properties:
        action:
          type: string
          enum:
          - HIDE
          - REJECT
          - SUSPEND
          - REQUEST_CORRECTION
          - REMOVE
          - SUSPEND_ACCOUNT
          - DISMISS
          - REVOKE_BADGE
          - ARCHIVE_PROFILE
          description: Closed verb set (BR-MOD-02).
        note:
          type: string
          maxLength: 1000
      additionalProperties: false
...
  examples:
    ErrValidation:
      summary: Field-level validation failure
      value:
        type: https://errors.activehome.uz/validation
        title: Validation failed
        status: 422
        code: VALIDATION_FAILED
        detail: One or more fields are invalid.
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a6b
        details:
        - path: /attributes/area_m2
          rule: numeric_range
          message: Must be between 1 and 100000.
        - path: /title
          rule: required
          message: Title is required.
    ErrBusinessRule:
      summary: Business-rule violation (quota)
      value:
        type: https://errors.activehome.uz/business-rule
        title: Listing quota exceeded
        status: 422
        code: QUOTA_EXCEEDED
        detail: Your current subscription allows 10 active listings.
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a6c
    ErrUnauthorized:
      summary: No valid session
      value:
        type: https://errors.activehome.uz/authentication
        title: Authentication required
        status: 401
        code: AUTHENTICATION_REQUIRED
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a6d
    ErrForbidden:
      summary: Permission denied
      value:
        type: https://errors.activehome.uz/authorization
        title: Permission denied
        status: 403
        code: PERMISSION_DENIED
        detail: You do not have permission to perform this action in the current acting context.
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a6e
    ErrNotFound:
      summary: Resource not found
      value:
        type: https://errors.activehome.uz/not-found
        title: Resource not found
        status: 404
        code: RESOURCE_NOT_FOUND
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a6f
    ErrConflict:
      summary: Optimistic-lock conflict
      value:
        type: https://errors.activehome.uz/conflict
        title: Update conflict
        status: 409
        code: CONFLICT
        detail: The resource was modified by another request. Re-fetch and retry.
        traceId: 01J9Z5K3QX8ac2f7b1e9d0c4a70
...